requirements.txt
Telethon~=1.39.0
python-dotenv~=1.0.1
colorlog~=6.9.0